new updates

+++ src/tasks/tasks.service.ts
@@ -14,14 +14,20 @@ export class TasksService {
     private taskRepository: TaskRepository,
   ){}
 
-async getTasks(filterDto: GetTasksFilterDto ): Promise<Task[]>{
-  return this.taskRepository.getTasks(filterDto);
+async getTasks(
+  filterDto: GetTasksFilterDto,
+  user: User, 
+  ): Promise<Task[]>{
+  return this.taskRepository.getTasks(filterDto, user);
 
 }
 
 
-async getTaskById(id: number): Promise<Task> {
-  const found = await this.taskRepository.findOne(id);
+async getTaskById(
+  id: number,
+  user: User,
+  ): Promise<Task> {
+  const found = await this.taskRepository.findOne({ where: { id, userId: user.id }});
     if (!found) {
    throw new NotFoundException(`Task with ID "${id}" not found`);
   }
@@ -33,13 +39,17 @@ async createTask(
   createTaskDto: CreateTaskDto,
   user: User,
   ): Promise<Task> {
-    
+
 
  return this.taskRepository.createTask(createTaskDto, user);  
 }
 
-async deleteTask(id: number) {
-const result = await this.taskRepository.delete(id);
+async deleteTask(
+ id: number,
+ user: User,
+ ): Promise<void> {
+
+const result = await this.taskRepository.delete({ id, userId: user.id });
 
 if( result.affected === 0){
   throw new NotFoundException(`Task with ID "${id}" not found`);
@@ -47,12 +57,17 @@ if( result.affected === 0){
 }
 }
 
-async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
- const task  = await this.getTaskById(id);
+async updateTaskStatus(
+  id: number, 
+  status: TaskStatus,
+  user: User,
+  ): Promise<Task> {
+ const task  = await this.getTaskById(id, user);
  task.status = status;
  await task.save();
  return task;
 }
+
 }
